Updates to header files and macros for RiRi project - Renamed the public header files from `.h` to `.hpp`. So, now public headers will be more easily recognized. - Private headers remain unchanged with `.h` extension. Making it easier to distinguish between public and private headers. - Added new header file `RiRiMacros.h` for common macros used in the project. - Added a better hash map implementation in `MemoryMaps.cpp` to improve performance and maintainability. - Updated `DataManager.h` to include the new macros and implemented data handling.
I am sorry for the lack of commits recently, I have been trying my best to come up with the best possible solution for the project (with procrastination, of course).
Additionally, I have been working on the documentation and examples to make it easier for users to understand how to use the library.

Finally, I apologize for the delay and a lack of more detailed commit message. I will cover these changes in detail in the next commit.