feat(updateValue): Implemented updateValue method in DataManager
- The `updateValue` method updates the value of a given key in the MemoryMap.
- It checks if the key exists and updates the value accordingly and returns true if successful.
- If the key does not exist, it returns false.

I know there is a `better looking` method:
```cpp
try {
    Internal::MemoryMap[key] = newValue;
    return true;
} catch (...) {
    return false;
}
```
And it's exactly what it's called, a `better looking` method.

In reality the `[]` creates a new entry if the key does not exist, which is not what we want here.
Plus, a `try-catch` block adds unnecessary overhead (it's zero cost, only if an exception is not thrown).

That is why I'd rather add O(1) complexity to the code and check if the key exists first, then update it, it may look slower, but it is the fastest available option.

Oh and if you thought:
```cpp
try {
    auto it = map.find(key);
    it->second = value;  // if key doesn't exist? -> doomsday
    return true;
} catch (...) {
    return false;
}
```
Yea, no, that `try-catch` isn't gonna save you regardless, because if the key doesn't exist, `it` will be an invalid iterator and dereferencing it will lead to undefined behavior. Which your favorite compiler `try-catch` block won't save you from.