fix(DataManager): Updated functions signatures to respect the `std::string` keys.
- `setValue()` now takes `std::string&&` type key parameter.
- Rest of the function signatures didn't require changes.

The `setValue()` implementation also now uses `std::move()` all the way from C++11 (old hag).

This, assumes that a `std::string` allocation of the key takes place somewhere in the upper layer function (command functions).
Then the `rvalue` of the key is passed down to the `setValue()` function.

Inside the `setValue()` function the key-values are inserted into the map using `.insert()`.

The value is passed as `RapidDataType` by reference and the key is passed in an interesting way:
```cpp
bool Internal::setValue(std::string&& key, const RapidDataType& value) noexcept {
    return Internal::MemoryMap.insert({std::move(key), value}).second;
}
```

Notice the key is first passed through `std::move()`, this makes it so that the key is now open to "getting owned" (I cannot phrase this better) by the `MemoryMap`.
And passing it to the `.insert()` function allows the `std::string` (which was originally allocated in the upper layer functions) to be moved into the map, avoiding unnecessary copies.
If you don't get it, just know that this allowed me to achieve 1 string allocation with `setValue()`. Which was the holy grail of this refactor.

1 allocation achieved. Crisis averted.

The PR will discuss the changes in detail.