feat(RapidResponse): migrate from .h to .hpp and basic implementation
- Changed RapidResponse from .h to .hpp to mark it as a public header. Even though it is inside the src/utils directory, it is intended for public use.
- Implemented basic functionality for RapidResponse, it's just a simple wrapper around the StatusCode enum.
- Implemented basic RapidResponseContainer, which is... heavy:

/--- RapidResponseContainer Explanation (I tried) ---/
If you know what Small Vector Optimization (SVO) is, then just know that we made a Array of Structs (structs are a pair of KEY(type: string_view) and StatusCode(type: enum)).
Then, we pre-allocated a block of memory and used it to store the RapidResponse objects by static allocation (typical SVO).

If that bounced off, or SVO doesn't ring a bell, then here is a more simplified explanation:
    - We pre-allocate a "block" or "blob" of memory (char[]) of apt size.
    - We then re-cast the char[] to a RapidResponse* array (dumb version: tells the compiler to treat the memory as an array of RapidResponse).
    - When needed, we add "KEY::StatusCode" pairs to the RapidResponse array using `new`, statically (essentially construct the objects inside the pre-allocated memory using pointer arithmetic).

This needs a LOT more detailing, but it is a good start. Expect a note in .notes shortly about this...
...with.. erm.. benchmarks, I do have them with me right now. Though I am a bit aghast at the results...
... it's in nanoseconds...what...nevermind that, that's my problem (or boon??) to deal with.

The names are temporary, they will be changed later (I think?) maybe the implementations too (slightly?).