note(>.>): The week long gap
There was a strange lack of commits for a week recently, and I must admit, I was not working on RiRi during this time. As clear by the recent commit which didn't have any groundbreaking changes.

This may seem very... inconsistent, but I had 'irl stuff' to deal with (driving tests, friends, a new... pet? a parakeet? etc.)

So, yes, I haven't abandoned RiRi and I don't plan to do so either, back to the usual 'almost' daily work on RiRi.

Also, uninstalling Visual Studio is a bad idea (this added... days to the gap); it screws up clang on windows. I do not know which individual components are required for clang to run on windows (with lld linker) yet. However, installing 'Desktop Development with C++' workload resolves it all (it has a lot of unwanted components, but that's the best I can tell now).