refactor(RapidTypes): Generalize `RapidCommandFn` and removed `RapidError.h`
- Replaced a specific ` RapidCommandFn ` return type with a templated `Response` type to ensure that it's compatible with different response types.
- Removed unused `RapidError.h` include.

Dev Notes:

We could have imported the "response types" from `RapidResponse.hpp`, but that file imports `RapidTypes.h` itself, creating a cycling import sadly, so we will just stick with a template for simplicity.

Honestly, I don't know if RapidCommandFn will even be useful anymore, it was made in the pre-restructure era of RiRi, but let's see.